Updated houseBuilder and Main

+++ src/main/kotlin/onm/house/devices/Oven.kt
@@ -14,7 +14,7 @@ class Oven(override val id: UUID,
 
     private val ovenBakeFinishedEvent = BakeFinishedEvent(eventHandler)
 
-    fun switchOn(food : Collection<Food>, minutes : Int){
+    fun switchOn(food: Collection<Food>, minutes: Double) {
 
         //TODO add execution in separate thread for $minutes minutes. Food size determines number of portions.
